+++ packages/algob/test/mocks/deployer.ts
@@ -19,6 +19,7 @@ export class FakeDeployer implements Deployer {
 	isDeployMode = false;
 	accounts = [];
 	accountsByName = new Map<string, rtypes.Account>();
+	assetPath = "assets";
 	scriptName = "";
 	checkpoint = {
 		getAppfromCPKey(key: string): rtypes.AppInfo | undefined {
@@ -118,7 +119,7 @@ export class FakeDeployer implements Deployer {
 		throw new Error("Not implemented");
 	}
 
-	addCheckpointKV(key: string, value: string): void { } // eslint-disable-line @typescript-eslint/no-empty-function
+	addCheckpointKV(key: string, value: string): void {} // eslint-disable-line @typescript-eslint/no-empty-function
 
 	getCheckpointKV(key: string): string | undefined {
 		return "metadata";